requirements.txt
wavelink
six
google_auth_oauthlib
discord.py
requests
discord
google_api_python_client
protobuf
python-dotenv
python_dateutil
